add method delete() that got lost in last merge delete-by-uuid
